Conflicts:
	src/Graxei/Graxei.Apresentacao.MVC4Unity/Controllers/HomeController.cs
	src/Graxei/Graxei.Apresentacao.MVC4Unity/Graxei.Apresentacao.MVC4Unity.csproj
	src/Graxei/Graxei.Persistencia.Implementacao.FluentNHibernate.Postgre/Graxei.Persistencia.Implementacao.FluentNHibernate.Postgre.csproj

+++ src/Graxei/Graxei.Aplicacao.Contrato/Consultas/IConsultaListaProdutosLoja.cs
@@ -9,6 +9,6 @@ namespace Graxei.Aplicacao.Contrato.Consultas
 {
     public interface IConsultaListaProdutosLoja
     {
-        ListaProdutosLoja Get(string criterio, bool meusProdutos, long idLoja, int pagina, int tamanhoPagina, int totalElementos);
+        ListaProdutosLoja Get(string criterio, bool meusProdutos, long idLoja, int pagina, int tamanhoPagina, long totalElementos);
     }
 }
